docs: change page files name
