--- frontend/src/views/pages/FriendsHistoryPage.tsx
@@ -1,177 +1,0 @@
-import { useState } from "react";
-import styled from "styled-components";
-import {
-  IconButton,
-  Toolbar,
-  Drawer,
-  useMediaQuery,
-  useTheme,
-  Box,
-} from "@mui/material";
-import MenuIcon from "@mui/icons-material/Menu";
-import { DrawerContents } from "../components/DrawerContents";
-import { GobackButton } from "../components/GobackButton";
-import { useNavigate } from "react-router-dom";
-
-export const FriendsHistoryPage = () => {
-  const [mobileOpen, setMobileOpen] = useState(false);
-  const materialTheme = useTheme();
-  const isMobile = useMediaQuery(materialTheme.breakpoints.down("sm"));
-
-  const handleDrawerToggle = () => {
-    setMobileOpen(!mobileOpen);
-  };
-
-  const navigate = useNavigate();
-
-  const handleGoBack = () => {
-    navigate("/history");
-  };
-
-  return (
-    <Wrapper>
-      <NavBox>
-        <DeskTopDrawer variant="permanent" open>
-          <Toolbar />
-          <DrawerContents />
-        </DeskTopDrawer>
-        <MobileDrawer
-          variant="temporary"
-          open={mobileOpen}
-          onClose={handleDrawerToggle}
-          ModalProps={{
-            keepMounted: true,
-          }}
-        >
-          <Toolbar />
-          <DrawerContents />
-        </MobileDrawer>
-      </NavBox>
-      <MainBox>
-        {isMobile && (
-          <StyledIconButton
-            color="inherit"
-            aria-label="open drawer"
-            edge="start"
-            onClick={handleDrawerToggle}
-          >
-            <MenuIcon />
-          </StyledIconButton>
-        )}
-        <SubBox>
-          <GobackButton onClick={handleGoBack} />
-          <DetailBox>
-            <Section>
-              <PageTitle>
-                Expense with<Span>Kota</Span>
-                <Span>Hana</Span>
-              </PageTitle>
-            </Section>
-            <TransactionCardContainer>
-              <Title>Summary for this group</Title>
-              <CalculateCardContainer></CalculateCardContainer>
-              <Title>All Expenses</Title>
-            </TransactionCardContainer>
-          </DetailBox>
-        </SubBox>
-      </MainBox>
-    </Wrapper>
-  );
-};
-
-const drawerWidth = 290;
-
-const Wrapper = styled.div`
-  display: flex;
-  height: calc(100% - 64px);
-`;
-
-const NavBox = styled.div`
-  flex-shrink: 0;
-
-  @media (min-width: 600px) {
-    width: ${drawerWidth}px;
-  }
-`;
-
-const DeskTopDrawer = styled(Drawer)`
-  display: none;
-  & .MuiDrawer-paper {
-    box-sizing: border-box;
-    width: ${drawerWidth}px;
-  }
-  @media (min-width: 600px) {
-    display: block;
-  }
-`;
-
-const MobileDrawer = styled(Drawer)`
-  display: block;
-  & .MuiDrawer-paper {
-    box-sizing: border-box;
-    width: ${drawerWidth}px;
-  }
-  @media (min-width: 600px) {
-    display: none;
-  }
-`;
-
-const StyledIconButton = styled(IconButton)`
-  .MuiSvgIcon-root {
-    color: ${({ theme }) => theme.palette.info.light};
-  }
-  .MuiTouchRipple-root {
-    color: ${({ theme }) => theme.palette.info.light};
-  }
-`;
-
-const MainBox = styled.div`
-  background-color: ${({ theme }) => theme.palette.primary.main};
-  padding: 50px 120px;
-  width: calc(100% - ${drawerWidth}px);
-  height: 100vh;
-  @media (max-width: 600px) {
-    width: 100%;
-    padding: 0px 20px;
-  }
-`;
-
-const PageTitle = styled.h2`
-  margin-top: 1rem;
-  padding-bottom: 2rem;
-  color: ${({ theme }) => theme.palette.info.light};
-`;
-
-const Title = styled.h3`
-  margin-top: 1rem;
-  margin-bottom: 1rem;
-  color: ${({ theme }) => theme.palette.info.light};
-`;
-
-const Section = styled.div`
-  width: 100%;
-  margin-bottom: 20px;
-`;
-
-const TransactionCardContainer = styled.div`
-  width: 100%;
-`;
-
-const DetailBox = styled.div`
-  margin: 0 1rem;
-`;
-
-const CalculateCardContainer = styled.div`
-  display: grid;
-  grid-template-columns: 1fr 1fr;
-  gap: 2rem;
-`;
-
-const Span = styled.span`
-  color: ${({ theme }) => theme.palette.secondary.light};
-  margin-left: 0.5rem;
-`;
-
-const SubBox = styled(Box)`
-  width: 100%;
-`;
