+++ frontend/src/views/pages/FriendsListPage.tsx
@@ -1,66 +1,47 @@
-import Box from "@mui/material/Box";
+import { useEffect, useState } from "react";
 import styled from "styled-components";
-import { createClient } from "@supabase/supabase-js";
-import { useCallback, useEffect, useState } from "react";
-import { useForm } from "react-hook-form";
-import { useSelector } from "react-redux";
-import { RootState } from "../../store/store";
-import { emailRegex } from "../../utils/regexPatternUtils";
-import { useNavigate } from "react-router-dom";
-import { InputAdornment, InputBase, Paper } from "@mui/material";
-import MailOutlineIcon from "@mui/icons-material/MailOutline";
-import { Friend, FriendWithStatus } from "../../types";
-import { SubButton } from "../components/SubButton";
-import { Database } from "../../../../supabase/schema";
-import { SupabaseEdgeFunctionService } from "../../services/supabaseEdgeFunction";
 import {
-  ERROR_EMAIL,
-  ERROR_SELECT_FRIENDS,
-  ERROR_SEND_EXISTED_ADDRESS,
-  ERROR_SEND_FAILED,
-  ERROR_SEND_OWN_ADDRESS,
-} from "../../constants/message";
-import { GobackButton } from "../components/GobackButton";
-import AccessTimeIcon from "@mui/icons-material/AccessTime";
+  IconButton,
+  Toolbar,
+  Drawer,
+  useMediaQuery,
+  useTheme,
+  Box,
+} from "@mui/material";
+import MenuIcon from "@mui/icons-material/Menu";
+import { DrawerContents } from "../components/DrawerContents";
+import { useSelector } from "react-redux";
+import { selectUser } from "../../reducer/userSlice";
+import { createClient } from "@supabase/supabase-js";
+import { FriendCard } from "../components/FriendCard";
+import { FriendApproveCard } from "../components/FriendApproveCard";
+import { FriendShipsReturns } from "./SelectFriendsPage";
 
-interface FriendEmail {
-  email: string;
-}
-
-const supabase = createClient<Database>(
+const supabase = createClient(
   process.env.REACT_APP_SUPABASE_URL as string,
   process.env.REACT_APP_SUPABASE_ANON_KEY as string
 );
 
-export type FriendShipsReturns =
-  Database["public"]["Functions"]["get_user_friends"]["Returns"];
-export type FriendShipsArgs =
-  Database["public"]["Functions"]["get_user_friends"]["Args"];
-export type FriendShipsInsert =
-  Database["public"]["Tables"]["Friendships"]["Insert"];
-
 export const FriendsListPage = () => {
-  const [selectedFriends, setSelectedFriends] = useState<Friend[]>([]);
-  const [friends, setFriends] = useState<FriendShipsReturns>([]);
-  const [selectedError, setSelectedError] = useState("");
+  const [mobileOpen, setMobileOpen] = useState(false);
+  const materialTheme = useTheme();
+  const isMobile = useMediaQuery(materialTheme.breakpoints.down("sm"));
   const [error, setError] = useState("");
-  const [success, setSuccess] = useState("");
-  const {
-    register,
-    handleSubmit,
-    formState: { errors },
-    reset,
-  } = useForm<FriendEmail>();
+  const [friends, setFriends] = useState<FriendShipsReturns>([]);
 
-  const userState = useSelector((state: RootState) => state.user);
-  const userId = userState.user?.id!;
-  const userEmail = userState.user?.email;
-  const userFirstName = userState.user?.firstName;
-  const userLastName = userState.user?.lastName;
+  const handleDrawerToggle = () => {
+    setMobileOpen(!mobileOpen);
+  };
 
-  const navigate = useNavigate();
+  const { user } = useSelector(selectUser);
+  const userId = user?.id;
 
-  const getUserFriendsById = useCallback(async () => {
+  useEffect(() => {
+    getUserFriendsById();
+    // eslint-disable-next-line react-hooks/exhaustive-deps
+  }, []);
+
+  const getUserFriendsById = async () => {
     try {
       const { data, error } = await supabase.rpc("get_user_friends", {
         user_id: userId,
@@ -70,410 +51,187 @@ export const FriendsListPage = () => {
         return false;
       } else {
         setFriends(data);
-        return true;
       }
     } catch (error: any) {
       setError(error.message);
       return false;
     }
-  }, [userId]);
-
-  useEffect(() => {
-    getUserFriendsById();
-  }, [getUserFriendsById]);
-
-  const onSubmit = (data: FriendEmail) => {
-    const { email } = data;
-    sendFriendRequest(email);
-  };
-
-  const sendFriendRequest = async (email: string) => {
-    const emailToLowerCase = email.toLowerCase();
-    if (emailToLowerCase === userEmail) {
-      setError(ERROR_SEND_OWN_ADDRESS);
-    } else {
-      const resultCountFriendShipByEmail = await countFriendShipByEmail(
-        emailToLowerCase
-      );
-      if (
-        typeof resultCountFriendShipByEmail === "number" &&
-        resultCountFriendShipByEmail > 0
-      ) {
-        setError(ERROR_SEND_EXISTED_ADDRESS);
-      } else {
-        const resultGetFriendByEmail = await getFriendByEmail(emailToLowerCase);
-        if (resultGetFriendByEmail) {
-          const resultInsertFriendship = await insertFriendship(
-            resultGetFriendByEmail,
-            emailToLowerCase
-          );
-          if (resultInsertFriendship) {
-            const emailResponse = await SupabaseEdgeFunctionService.sendEmail(
-              email,
-              `${userFirstName} ${userLastName}`
-            );
-
-            if (!emailResponse.status) {
-              setError(ERROR_SEND_FAILED);
-            }
-
-            // to retrieve the data to update the friend list
-            const resultGetUserFriendsById = await getUserFriendsById();
-            if (resultGetUserFriendsById) {
-              reset();
-              setError("");
-            }
-            setSuccess(
-              `You have successfully sent a friend request to ${email}!`
-            );
-          }
-        }
-      }
-    }
-  };
-
-  const handleSendEmail = () => {
-    console.log("send mail");
-  };
-
-  // Check friends to add or not
-  const handleCheckedChange = (
-    id: string,
-    email: string,
-    firstName: string,
-    lastName: string,
-    isChecked: boolean
-  ): void => {
-    if (isChecked) {
-      const addFriend: Friend = {
-        id: id,
-        firstName: firstName,
-        lastName: lastName,
-        email: email,
-      };
-      setSelectedFriends((prevSelectedFriends) => [
-        ...prevSelectedFriends,
-        addFriend,
-      ]);
-    } else {
-      setSelectedFriends((prevSelectedFriends) =>
-        prevSelectedFriends.filter((person) => person.email !== email)
-      );
-    }
-  };
-
-  // check if a user has already sent a friend request to the input email address
-  const countFriendShipByEmail = async (email: string) => {
-    try {
-      const { data, error } = await supabase.rpc("check_friendship", {
-        user_id: userId,
-        friend_email: email,
-      });
-      if (error) {
-        setError(error.message);
-        return false;
-      } else {
-        return data;
-      }
-    } catch (error: any) {
-      setError(error.message);
-      return false;
-    }
-  };
-
-  const getFriendByEmail = async (email: string) => {
-    try {
-      const { data, error } = await supabase
-        .from("Users")
-        .select("*")
-        .eq("email", email);
-      if (error) {
-        setError(error.message);
-        return false;
-      } else {
-        return data;
-      }
-    } catch (error: any) {
-      setError(error.message);
-      return false;
-    }
-  };
-
-  // register the input email address to friendships table for a friend request
-  const insertFriendship = async (data: any, email: string) => {
-    const friendshipsData: FriendShipsInsert = {
-      userId: userId,
-      friendId: data.length > 0 ? data[0].id : null,
-      friendEmail: data.length > 0 ? data[0].email : email,
-      statusId: 1, // status: pending
-      registeredAt: new Date().toISOString(),
-      updatedAt: new Date().toISOString(),
-    };
-
-    try {
-      const { error } = await supabase
-        .from("Friendships")
-        .insert(friendshipsData);
-      if (error) {
-        setError(error.message);
-        return false;
-      } else {
-        return true;
-      }
-    } catch (error: any) {
-      setError(error.message);
-      return false;
-    }
-  };
-
-  const handleClick = () => {
-    if (selectedFriends.length > 0) {
-      setSelectedError("");
-      navigate("/expenses/payment", {
-        state: { selectedFriends },
-      });
-    } else {
-      setSelectedError(ERROR_SELECT_FRIENDS);
-    }
-  };
-
-  const handleGoBack = () => {
-    navigate("/history");
   };
 
   return (
-    <MainContainer>
-      <SubContainer>
-        <GobackButtonWrapper>
-          <GobackButton onClick={handleGoBack} />
-        </GobackButtonWrapper>
-        <Title>Add new friends</Title>
-        <StyledBox component="form" onSubmit={handleSubmit(onSubmit)}>
-          <InputWrapper>
-            <InputPaper elevation={0}>
-              <InputAdornment position="start">
-                <MailOutlineIcon />
-              </InputAdornment>
-              <InputBase
-                placeholder="Email"
-                type="email"
-                fullWidth
-                {...register("email", {
-                  required: true,
-                  pattern: emailRegex,
-                })}
-              />
-            </InputPaper>
-            {errors.email && <ErrorText>{ERROR_EMAIL}</ErrorText>}
-          </InputWrapper>
-          <SubButtonWrapper>
-            <SubButton title={"send"} onClick={handleSendEmail} />
-            {error ? (
-              <ErrorText>{error}</ErrorText>
-            ) : success ? (
-              <SuccessText>{success}</SuccessText>
-            ) : null}
-          </SubButtonWrapper>
-        </StyledBox>
-      </SubContainer>
-      <FriendListSubContainer>
-        <Title>Friendslist</Title>
-        <UnorderedList>
-          {/* statusId 1: pending, 2: approved */}
-          {/* friend.id could be null so pathing index instead of friend.id */}
-          {friends!
-            .filter((friend) => friend.statusId === 2 || friend.sender)
-            .map((friend: FriendWithStatus, index) => {
-              return (
-                <List key={index}>
-                  <CheckBox
-                    type="checkbox"
-                    id={index.toString()}
-                    checked={
-                      !!selectedFriends.find(
-                        (selectedFriend) =>
-                          selectedFriend.email === friend.email
-                      )
-                    }
-                    onChange={(event) =>
-                      handleCheckedChange(
-                        friend.id,
-                        friend.email,
-                        friend.firstName,
-                        friend.lastName,
-                        event.target.checked
-                      )
-                    }
-                    disabled={friend.statusId === 1}
-                  />
-                  <Label htmlFor={index.toString()}>
-                    {friend.statusId !== 1 ? (
-                      <ListItem>
-                        {friend.firstName} {friend.lastName}
-                      </ListItem>
-                    ) : (
-                      <PendingFriendList>
-                        <AccessTimeIcon fontSize="small" />
-                        <PendingListItem>pending approval</PendingListItem>
-                      </PendingFriendList>
+    <Wrapper>
+      <NavBox>
+        <DeskTopDrawer variant="permanent" open>
+          <Toolbar />
+          <DrawerContents />
+        </DeskTopDrawer>
+        <MobileDrawer
+          variant="temporary"
+          open={mobileOpen}
+          onClose={handleDrawerToggle}
+          ModalProps={{
+            keepMounted: true,
+          }}
+        >
+          <Toolbar />
+          <DrawerContents />
+        </MobileDrawer>
+      </NavBox>
+      <MainBox>
+        {isMobile && (
+          <StyledIconButton
+            color="inherit"
+            aria-label="open drawer"
+            edge="start"
+            onClick={handleDrawerToggle}
+          >
+            <MenuIcon />
+          </StyledIconButton>
+        )}
+        <SubBox>
+          <DetailBox>
+            <Section>
+              <Title>Your Friends List</Title>
+              {friends!.some(
+                (friend) => !friend.sender && friend.statusId === 1 && friend.id
+              ) ? (
+                <>
+                  <SubTitle>Friend Request</SubTitle>
+                  <hr />
+                  <UnorderedList>
+                    {friends!.map((friend, index) =>
+                      !friend.sender && friend.statusId === 1 && friend.id ? (
+                        <FriendApproveCard
+                          id={friend.id}
+                          firstName={friend.firstName}
+                          lastName={friend.lastName}
+                          email={friend.email}
+                          key={index}
+                          getUserFriendsById={getUserFriendsById}
+                        />
+                      ) : null
                     )}
-                    <ListItem>{friend.email}</ListItem>
-                  </Label>
-                </List>
-              );
-            })}
-        </UnorderedList>
-        <ButtonContainer>
-          <SubButton title={"create"} onClick={handleClick} />
-        </ButtonContainer>
-        {selectedFriends.length === 0 && <ErrorText>{selectedError}</ErrorText>}
-      </FriendListSubContainer>
-    </MainContainer>
+                  </UnorderedList>
+                </>
+              ) : null}
+              <SubTitle>All friends</SubTitle>
+              <hr />
+              <UnorderedList>
+                {friends!.map((friend, index) =>
+                  friend.statusId === 2 ? (
+                    <FriendCard
+                      id={friend.id}
+                      firstName={friend.firstName}
+                      lastName={friend.lastName}
+                      email={friend.email}
+                      key={index}
+                    />
+                  ) : null
+                )}
+              </UnorderedList>
+              {error && <ErrorText>{error}</ErrorText>}
+            </Section>
+          </DetailBox>
+        </SubBox>
+      </MainBox>
+    </Wrapper>
   );
 };
 
-const MainContainer = styled.div`
-  height: calc(100% - 64px);
-  width: 100%;
-  overflow: auto;
+const drawerWidth = 290;
+
+const Wrapper = styled.div`
   display: flex;
-  justify-content: center;
-  gap: 20px;
-  background: ${({ theme }) => theme.palette.primary.main};
+  height: calc(100% - 64px);
+`;
+
+const NavBox = styled.div`
+  flex-shrink: 0;
+
+  @media (min-width: 600px) {
+    width: ${drawerWidth}px;
+  }
+`;
+
+const DeskTopDrawer = styled(Drawer)`
+  display: none;
+  & .MuiDrawer-paper {
+    box-sizing: border-box;
+    width: ${drawerWidth}px;
+  }
+
+  @media (min-width: 600px) {
+    display: block;
+  }
+`;
+
+const MobileDrawer = styled(Drawer)`
+  display: block;
+  & .MuiDrawer-paper {
+    box-sizing: border-box;
+    width: ${drawerWidth}px;
+  }
+  @media (min-width: 600px) {
+    display: none;
+  }
+`;
+
+const StyledIconButton = styled(IconButton)`
+  .MuiSvgIcon-root {
+    color: ${({ theme }) => theme.palette.info.light};
+  }
+  .MuiTouchRipple-root {
+    color: ${({ theme }) => theme.palette.info.light};
+  }
+`;
+
+const MainBox = styled.div`
+  background-color: ${({ theme }) => theme.palette.primary.main};
+  padding: 50px 120px;
+  width: calc(100% - ${drawerWidth}px);
   height: 100vh;
   @media (max-width: 600px) {
-    flex-direction: column;
-    margin-top: 0;
-  }
-`;
-
-const FriendListSubContainer = styled(Box)`
-  display: flex;
-  flex-direction: column;
-  align-items: center;
-  height: 70%;
-  width: 35%;
-  padding: 2rem;
-  background: ${({ theme }) => theme.palette.primary.main};
-  margin-top: 70px;
-  @media (max-width: 600px) {
-    height: 100%;
     width: 100%;
-    padding: 0;
+    padding: 0 20px;
   }
 `;
 
-const SubContainer = styled(Box)`
-  display: flex;
-  flex-direction: column;
-  align-items: center;
-  height: 70%;
-  width: 35%;
-  padding: 2rem;
-  background: ${({ theme }) => theme.palette.primary.main};
-  margin-top: 30px;
-  @media (max-width: 600px) {
-    height: 100%;
-    width: 100%;
-    padding: 0;
-  }
+const Section = styled.div`
+  width: 100%;
+  margin-bottom: 20px;
 `;
 
-const Title = styled.h1`
-  margin-top: 1rem;
-  margin-bottom: 1rem;
-  color: ${({ theme }) => theme.palette.secondary.main};
+const DetailBox = styled.div`
+  margin: 0 1rem;
+`;
+
+const SubBox = styled(Box)`
+  width: 100%;
 `;
 
 const UnorderedList = styled.ul`
   padding: 0px;
-  height: 60%;
-  width: 80%;
-  overflow: auto;
+  height: 100%;
+  width: 100%;
 `;
 
-const List = styled.li``;
-
-const CheckBox = styled.input`
-  display: none;
+const Title = styled.h2`
+  margin-top: 1rem;
+  padding-bottom: 2rem;
+  color: ${({ theme }) => theme.palette.info.light};
 `;
 
-const Label = styled.label`
-  display: flex;
-  justify-content: space-between;
-  border: 1px solid ${({ theme }) => theme.palette.secondary.main};
-  border-radius: 0.25rem;
-  padding: 0.6rem;
-  margin-bottom: 0.75rem;
-  color: ${({ theme }) => theme.palette.secondary.main};
-  cursor: pointer;
-  ${CheckBox}:checked + & {
-    background: #e0e0e0;
-  }
-`;
-
-const ListItem = styled.span`
-  display: inline-block;
-`;
-
-const PendingFriendList = styled.div`
-  display: flex;
-  flex-direction: row;
-  align-items: center;
-`;
-
-const PendingListItem = styled.span`
-  display: inline-block;
-  padding: 0 0.4rem;
-  font-size: 0.8rem;
-`;
-
-const ButtonContainer = styled.div`
-  display: flex;
-  width: 80%;
+const SubTitle = styled.h3`
+  margin-top: 1rem;
+  margin-bottom: 1rem;
+  color: ${({ theme }) => theme.palette.info.light};
 `;
 
 const ErrorText = styled.div`
+  display: flex;
+  justify-content: center;
+  width: 100%;
   margin-top: 7px;
   font-size: 1rem;
   color: #ff908d;
-`;
-
-const SuccessText = styled.div`
-  margin-top: 7px;
-  font-size: 1rem;
-  color: #4caf50;
-`;
-
-const InputWrapper = styled.div`
-  width: 100%;
-  display: flex;
-  flex-direction: column;
-`;
-
-const InputPaper = styled(Paper)`
-  margin: 15px 0 0;
-  padding: 7px;
-  display: flex;
-  align-items: center;
-  border: solid 1px ${({ theme }) => theme.palette.secondary.main};
-`;
-
-const SubButtonWrapper = styled.div`
-  width: 100%;
-  display: flex;
-  flex-direction: column;
-  margin-top: 30px;
-`;
-
-const StyledBox = styled(Box)`
-  display: flex;
-  flex-direction: column;
-  width: 70%;
-  margintop: 2rem;
-`;
-
-const GobackButtonWrapper = styled.div`
-  width: 70%;
-  display: flex;
 `;
